requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
transformers==4.35.2
torch==2.8.0
sentence-transformers==2.2.2
wikipedia==1.4.0
chromadb==0.4.15
pydantic==2.5.0
numpy==1.24.3
python-multipart==0.0.6